remove hold header
