--- src/presentation/components/Header/types.ts
@@ -1,8 +1,0 @@
-export type IHeaderProps = {
-  scrollIntoView: () => void;
-  className: string;
-  onOpenSearchModal?: () => void;
-  onResetSearch?: () => void;
-  openMobileMenu: boolean;
-  setOpenMobileMenu: (a: boolean) => void;
-};
